Add play methods to audio Output
